Added new `ignoreSleepingPercentage` option + implementation. Renamed config to offline_config so it didn't clash with TwinSessions when debugging. Updated changelog Updated readme
