+++ gradle.properties
@@ -8,7 +8,7 @@ minecraft_version=1.21.1
 loader_version=0.15.11
 
 # Mod Properties
-mod_version=1.1.0
+mod_version=1.1.1
 maven_group=com.misterfish
 archives_base_name=offlineplayersreworked
