+++ src/main/java/com/misterfish/commands/OfflinePlayerCommands.java
@@ -1,7 +1,7 @@
 package com.misterfish.commands;
 
 import com.misterfish.OfflinePlayersReworked;
-import com.misterfish.config.ModConfigs;
+import com.misterfish.offline_config.ModConfigs;
 import com.misterfish.exception.InvalidActionException;
 import com.misterfish.exception.InvalidIntervalException;
 import com.misterfish.exception.InvalidOffsetException;
